Release class Preprocessor with secondary function
The class recognizes directives specified by the user in the program text and substitutes the text of the found file in their place. In the initial version of the preprocessor, only the "include" directive with quotes "" and brackets <> is recognized. The predicate auxiliary function checks for the presence of a file among a vector of directories.